+++ client/cl_view.c
@@ -368,9 +368,9 @@ float CalcFov (float fov_x, float width, float height)
 	if (fov_x < 1 || fov_x > 179)
 		Com_Error (ERR_DROP, "Bad fov: %f", fov_x);
 
-	x = width/tan(fov_x/360*Q_PI);
+	x = width/tanf(fov_x/360*Q_PI);
 
-	a = atan (height/x);
+	a = atanf (height/x);
 
 	a = a*360/Q_PI;
